+++ integration_alg_gpu.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-float gpu_integrate(float *x_list, float *y_list, int length);
+float gpu_integrate(float *x_list, float *y_list, long long length);
 
 #endif
